requirements.txt
Babel==2.5.1
docopt==0.6.2
et-xmlfile==1.0.1
jdcal==1.3
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
pytz==2017.3
records==0.5.2
SQLAlchemy==1.1.14
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
numpy==1.5.0
mxnet-cu101==1.7.0
